% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsRunUtilities.R
\name{addNewRun2DB}
\alias{addNewRun2DB}
\title{Add the output to the master FVSOut.db}
\usage{
addNewRun2DB(runuuid, dbcon)
}
\arguments{
\item{runuuid}{a character string with the run UUID}

\item{dbcon}{the database connection to FVSOut.db}
}
\value{
a character string with an informative message as to what was done.
}
\description{
Add the output from a recently completed run to the master output database,
FVSOut.db
}
