% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalCallable.R
\name{externalDeleteComponents}
\alias{externalDeleteComponents}
\title{Delete FVS keyword components from a run.}
\usage{
externalDeleteComponents(prjDir = getwd(), runUUID, compUUIDs)
}
\arguments{
\item{prjDir}{is the path name to the project directory, if null the 
current directory is the project directory.}

\item{runUUID}{a character string of the run uuid that is processed}

\item{compUUIDs}{a vector of character strings holding the UUIDs of
the components that will be deleted from the run.}
}
\value{
the number of deletions.
}
\description{
Given a project directory a run uuid, and a list of component UUIDs,
this function deletes those components.
}
