% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalCallable.R
\name{externalDeleteRuns}
\alias{externalDeleteRuns}
\title{Delete FVS Runs}
\usage{
externalDeleteRuns(prjDir = NULL, runUUIDs = NULL, delOutput = TRUE)
}
\arguments{
\item{prjDir}{is the path name to the project directory, if null the 
current directory is the project directory.}

\item{runUUIDs}{a character vector of 1 or more run uuids to be deleted.}

\item{delOutput}{if TRUE (the default) the data in FVSOut.db is also
deleted.}
}
\value{
a data.frame listing the uuid, title, and datetime of the remaining runs
  and NULL if the no runs exist.
}
\description{
Given a project directory and 1 or more run uuids, the function deletes
the runs.
}
