% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalCallable.R
\name{externalDuplicateRun}
\alias{externalDuplicateRun}
\title{Duplicate a run and give the duplicate a new title}
\usage{
externalDuplicateRun(
  prjDir = getwd(),
  runUUID = NULL,
  dupTitle = NULL,
  dupMgmtID = NULL
)
}
\arguments{
\item{prjDir}{is the path name to the project directory, if null the 
current directory is the project directory.}

\item{runUUID}{the uuid of the run that will be duplicated.}

\item{dupTitle}{a character string with the duplicated run's title
if null, the system generates the name.}

\item{dupMgmtID}{a character string with the duplicated run's management ID.
if null, the system generates the ID.}
}
\value{
the new run uuid, NULL if not created.
}
\description{
Pass in a project directory and an existing runUUID, the duplicates title
and the run is duplicated and given the new title and default management ID.
}
