% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalCallable.R
\name{externalGetRuns}
\alias{externalGetRuns}
\title{Get FVS Runs}
\usage{
externalGetRuns(prjDir = getwd())
}
\arguments{
\item{prjDir}{is the path name to the project directory, default is
current directory.}
}
\value{
a data.frame listing the uuid, title, and datetime of existing runs
  and NULL if the project does not exist or if no runs exist.
}
\description{
Pass in a project directory and get back a data.fram of the FVS runs.
}
