% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalCallable.R
\name{externalMakeRun}
\alias{externalMakeRun}
\title{Build an FVS run in a project}
\usage{
externalMakeRun(
  prjDir = getwd(),
  title = NULL,
  standIDs = NULL,
  stdInit = "FVS_StandInit",
  variant
)
}
\arguments{
\item{prjDir}{is the path name to the project directory, if null the system
assumes that the current directory is the project directory.}

\item{title}{a character string with the run title, if null, the system generates 
the name, that is, it would be "Run 1", or "Run 2", and so on for 
all the runs already present. If none are present, then the default 
run title with be "Run 1" corresponding to the first run.}

\item{standIDs}{a vector of character strings holding the stand IDs (for example: 
\code{standIDs=c("id1","id2")} would load 2 stands, whereas: \code{standIDs="id1"} 
 would load just one. If NULL, all stands in the init table are loaded.}

\item{stdInit}{a character string of the name of the standinit table you 
want to use, the stands would be loaded from that table.}

\item{variant}{a 2 character string specifying the variant (required).
Ihe standID and the variant must match in the designated init table.}

\item{autoOut}{a character vector of output table selections using these
codes (if null, defaults are used):}

\item{startyr}{the start year of the simulation, if NULL, the current year
is used. a 2 character string specifying the variant (required).}

\item{endyr}{the end year of the simulation, if NULL one is computed using
variant-specific settings.}

\item{cycleat}{a vector of years where cycle boundaries are requested, none
outside the startyr to endyr interval are used.}
}
\value{
the new run uuid, NULL if not created.
}
\description{
Build an FVS run in a project and add it to the list of runs in the project. 
If some FVS runs are already present, another is added with this call.
The name of the input data base is FVS_Data.db and it must already exist.
}
