% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalCallable.R
\name{externalSetSomeOptions}
\alias{externalSetSomeOptions}
\title{Set FVS output and timing options}
\usage{
externalSetSomeOptions(
  prjDir = getwd(),
  runUUID,
  autoOut = NULL,
  svsOut = NULL,
  startyr = NULL,
  endyr = NULL,
  cyclelen = NULL,
  cycleat = NULL
)
}
\arguments{
\item{prjDir}{is the path name to the project directory, if null the 
current directory is the project directory.}

\item{runUUID}{a character vector of the run uuid that is processed}

\item{autoOut}{a vector of character strings corresponding to the automatic
output selections to be set (not the svs ones), where these values "turn on"
the corresponding selections and existing selections are not changed (if NULL
no changes are made: "Treelists", "Carbon", "Fire", "Dead", "CanProfile", 
"SnagDet",  "StrClass", "CalibStats", "Climate", "Econ", "DM_Sz_Sum", 
"RD_Sum",  "RD_Det", "RD_Beetle", "InvStats", "Regen", "KeepTextTables"}

\item{svsOut}{a vector of two character strings where the first is the
plot shape ("round" or "square") and the second is the number of images
of firelines, for example: svsOut=c("square",2). If only one value is given
it must be the plot shape and the number of intervals is set to 4.}

\item{startyr}{is the common starting year of the simulation, if NULL, 
no changes are made.}

\item{endyr}{is the common ending year of the simulation, if NULL, 
no changes are made.}

\item{cyclelen}{the "normal" cycle length, usually left to the variant-specific default.}

\item{cycleat}{is a vector of years where cycles are scheduled, if NULL, 
no changes are made.}
}
\value{
the runUUID if changes are made, else NULL
}
\description{
Given a project directory a run uuid, set the output selections
and the timing keyword controls
}
