% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{fvsOL}
\alias{fvsOL}
\title{Run fvsOL (FVS OnLine/OnLocal).}
\usage{
fvsOL(prjDir = NULL, fvsBin = NULL, shiny.trace = FALSE)
}
\arguments{
\item{prjDir}{the name of the directory containing an fvsOL project.}

\item{fvsBin}{the name of the directory containing the FVS load libraries for the platform}

\item{shiny.trace}{turns on tracing for shiny, see shiny documentation}
}
\value{
the shiny app.
}
\description{
Run fvsOL (FVS OnLine/OnLocal).
}
